Move shared cloud packages
