+++ bootstrapper/internal/kubernetes/kubernetes_test.go
@@ -31,9 +31,7 @@ import (
 )
 
 func TestMain(m *testing.M) {
-	goleak.VerifyTestMain(m,
-		goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"),
-	)
+	goleak.VerifyTestMain(m)
 }
 
 func TestInitCluster(t *testing.T) {
