--- internal/azureshared/doc.go
@@ -1,16 +1,0 @@
-/*
-Copyright (c) Edgeless Systems GmbH
-
-SPDX-License-Identifier: AGPL-3.0-only
-*/
-
-package azureshared
-
-/*
-Package azureshared contains code that is related to Microsoft Azure
-and is used by multiple microservices.
-
-This package is intended to have a minimal size and surface. If you
-have Azure related code that is not shared by multiple microservices,
-please keep the code in the microservice's internal package.
-*/
