+++ internal/cloud/azureshared/appcredentials.go
