+++ internal/cloud/azureshared/appcredentials_test.go
