+++ internal/cloud/azureshared/doc.go
@@ -1,0 +1,16 @@
+/*
+Copyright (c) Edgeless Systems GmbH
+
+SPDX-License-Identifier: AGPL-3.0-only
+*/
+
+/*
+Package gcpshared contains code to parse and define data types
+relevant for Microsoft Azure.
+
+This package is intended to have a minimal size and surface. If you
+have Azure related code that is not shared by multiple applications,
+or if the code interacts with the GCP API, please keep the code in
+the application's internal package or add it to the Azure cloud package.
+*/
+package azureshared
