+++ internal/cloud/azureshared/metadata.go
