+++ internal/cloud/azureshared/metadata_test.go
