+++ internal/cloud/gcpshared/doc.go
@@ -1,0 +1,15 @@
+/*
+Copyright (c) Edgeless Systems GmbH
+SPDX-License-Identifier: AGPL-3.0-only
+*/
+
+/*
+Package gcpshared contains code to parse and define data types
+relevant for Google Cloud Platform.
+
+This package is intended to have a minimal size and surface. If you
+have GCP related code that is not shared by multiple applications,
+or if the code interacts with the GCP API, please keep the code in
+the application's internal package or add it to the GCP cloud package.
+*/
+package gcpshared
