+++ internal/cloud/gcpshared/serviceaccountkey.go
@@ -4,7 +4,7 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package gcp
+package gcpshared
 
 import (
 	"fmt"
